// src/Api/merchant/dto/merchant.dto.ts
export class UserMerchantDto {
  _id: string;
  readonly user: string;
  readonly psp: string;
  ref: string;
  merchantcode: any;
  readonly title: string;
  readonly address: string;
  readonly province: string;
  readonly city: string;
  readonly tell: string;
  readonly status: boolean;
  readonly category: string;
  readonly email: string;
  readonly lat: number;
  readonly long: number;
  logo: string;
  img1: string;
  img2: string;
  img3: string;
}