// src/Api/merchant/dto/newterminal.dto.ts
export class MerchantTerminalDto {
  user: string;
  psp: string;
  merchant?: string;
  title: string;
  address: string;
  province: string;
  city: string;
  terminalid: number;
  status: boolean;
  sheba: string;
  cosheba: string;
  point?: string;
  discnonebank: number;
  discbank: number;
  wage?: string;
  rate?: number;
  visit?: number;
  mobile?: number;
  fullname?: string;
  nationalcode: string;
  password?: string;
  type?: string;
}